backend/requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
pymongo==4.5.0
requests==2.31.0
python-dotenv==1.0.0
motor==3.3.1
schedule==1.2.0
pandas==2.2.3